Moving euphonic tests to their own folder
